in-class/5_proof_of_django/TokenAuth/requirements.txt
django==5.0.7
channels==4.1.0
django-rest-framework==0.1.0
django_cors_headers==4.4.0